+++ frontend/src/features/customers/filters/FilterDialog.tsx
@@ -91,25 +91,14 @@ const FilterDialog: React.FC<FilterDialogProps> = ({
   };
 
   const handleSaveFilter = async () => {
-    if (selectedFilter) {
-      // Update existing filter
-      const updatedFilter: SavedFilter = {
-        ...selectedFilter,
+    const name = prompt('Enter a name for this filter:');
+    if (name) {
+      const newFilter: SavedFilter = {
+        id: crypto.randomUUID(),
+        name,
         filter: currentFilter,
       };
-      await dispatch(saveFilter(updatedFilter));
-      handleMenuClose();
-    } else {
-      // Create new filter
-      const name = prompt('Enter a name for this filter:');
-      if (name) {
-        const newFilter: SavedFilter = {
-          id: crypto.randomUUID(),
-          name,
-          filter: currentFilter,
-        };
-        await dispatch(saveFilter(newFilter));
-      }
+      await dispatch(saveFilter(newFilter));
     }
   };
 
@@ -176,10 +165,6 @@ const FilterDialog: React.FC<FilterDialogProps> = ({
           open={Boolean(menuAnchor)}
           onClose={handleMenuClose}
         >
-          <MenuItem onClick={handleSaveFilter}>
-            <SaveIcon sx={{ mr: 1 }} />
-            Update Filter
-          </MenuItem>
           <MenuItem onClick={handleDeleteFilter}>
             <DeleteIcon sx={{ mr: 1 }} />
             Delete Filter